add purpose of processing and CBP URL to the CBP model
There is no way to translate the registration number to a valid CBP URL. That
is why we save the url. at some point we want to change this to just the unique
internal cbp identifier (moid). Or tell the CBP their application is retarted,
and they should fix it.