Introduce the new filtering system.
Currently it has a load of duplicate code. It isn't clear yet which way it will
go, so abstracting it at this point doesn't seem like a good idea to me.

However, the code is quite hard to manage like this, so a new form is needed
soon when it's clear which way we'll be going.